requirements.txt
fastapi==0.95.1
cassandra-driver==3.26.0
uvicorn==0.21.1
python_dotenv==1.0.0
